fix: remove unused docs / private type
